chore: refactor controller
